#88 Setup Dynamic Value Asignement
