+++ Frontend/src/app/app.component.html
@@ -1,5 +1,3 @@
 <app-nav-bar></app-nav-bar>
 
 <router-outlet></router-outlet>
-
-<button (click)="testDL()">DownLoad Test</button>
